requirements-dev.txt
-r requirements.txt
black==24.8.0
ruff==0.5.7
mypy==1.11.1
pytest==8.3.2
typing-extensions==4.12.2
pandas-stubs==2.2.2.240807
pydantic-settings==2.4.0
pytest-cov==5.0.0
pylint
coverage
xenon
bandit
semgrep
pip-audit
deptry
